\name{futils-package}
\alias{futils-package}
\alias{futils}
\docType{package}
\title{
Farid's utility functions
}
\description{

}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
Farid Cheraghi

Maintainer: Farid Cheraghi
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = futils)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{futils}}
%\section{Documentation Index}{\packageIndices{futils}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
